+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 12.x.x (unreleased)
 
 ### Improvements
+- Client: Preferences Registry
 
 ### Bugfixes
 - 
@@ -134,8 +135,6 @@ same as 12.112.0
 - Server: Improve performance
 - Server: Supports IPv6 on Redis transport. @mei23  
   IPv4/IPv6 is used by default. You can tune this behavior via `redis.family`.
-- Migrate to Yarn Berry (v3.2.1) @ThatOneCalculator
-	- You may have to `yarn run clean-all` and `yarn set version berry` before running `yarn install` if you're still on yarn classic
 - Server: Add possibility to log IP addresses of users @syuilo
 - Add additional drive capacity change support @CyberRex0